Refactored  process and Python environment management modules
