+++ core/process_manager.py
@@ -1,0 +1,139 @@
+# Process management for command termination
+import threading
+import time
+import logging
+import os
+import signal
+
+active_processes = {}  # pid -> process info
+process_lock = threading.Lock()
+logger = logging.getLogger(__name__)
+
+class ProcessManager:
+    """Enhanced process manager for command termination and monitoring"""
+
+    @staticmethod
+    def register_process(pid, command, process_obj):
+        """Register a new active process"""
+        with process_lock:
+            active_processes[pid] = {
+                "pid": pid,
+                "command": command,
+                "process": process_obj,
+                "start_time": time.time(),
+                "status": "running",
+                "progress": 0.0,
+                "last_output": "",
+                "bytes_processed": 0
+            }
+            logger.info(f"🆔 REGISTERED: Process {pid} - {command[:50]}...")
+
+    @staticmethod
+    def update_process_progress(pid, progress, last_output="", bytes_processed=0):
+        """Update process progress and stats"""
+        with process_lock:
+            if pid in active_processes:
+                active_processes[pid]["progress"] = progress
+                active_processes[pid]["last_output"] = last_output
+                active_processes[pid]["bytes_processed"] = bytes_processed
+                runtime = time.time() - active_processes[pid]["start_time"]
+
+                # Calculate ETA if progress > 0
+                eta = 0
+                if progress > 0:
+                    eta = (runtime / progress) * (1.0 - progress)
+
+                active_processes[pid]["runtime"] = runtime
+                active_processes[pid]["eta"] = eta
+
+    @staticmethod
+    def terminate_process(pid):
+        """Terminate a specific process"""
+        with process_lock:
+            if pid in active_processes:
+                process_info = active_processes[pid]
+                try:
+                    process_obj = process_info["process"]
+                    if process_obj and process_obj.poll() is None:
+                        process_obj.terminate()
+                        time.sleep(1)  # Give it a chance to terminate gracefully
+                        if process_obj.poll() is None:
+                            process_obj.kill()  # Force kill if still running
+
+                        active_processes[pid]["status"] = "terminated"
+                        logger.warning(f"🛑 TERMINATED: Process {pid} - {process_info['command'][:50]}...")
+                        return True
+                except Exception as e:
+                    logger.error(f"💥 Error terminating process {pid}: {str(e)}")
+                    return False
+            return False
+
+    @staticmethod
+    def cleanup_process(pid):
+        """Remove process from active registry"""
+        with process_lock:
+            if pid in active_processes:
+                process_info = active_processes.pop(pid)
+                logger.info(f"🧹 CLEANUP: Process {pid} removed from registry")
+                return process_info
+            return None
+
+    @staticmethod
+    def get_process_status(pid):
+        """Get status of a specific process"""
+        with process_lock:
+            return active_processes.get(pid, None)
+
+    @staticmethod
+    def list_active_processes():
+        """List all active processes"""
+        with process_lock:
+            return dict(active_processes)
+
+    @staticmethod
+    def pause_process(pid):
+        """Pause a specific process (SIGSTOP)"""
+        with process_lock:
+            if pid in active_processes:
+                try:
+                    process_obj = active_processes[pid]["process"]
+                    if process_obj and process_obj.poll() is None:
+                        os.kill(pid, signal.SIGSTOP)
+                        active_processes[pid]["status"] = "paused"
+                        logger.info(f"⏸️  PAUSED: Process {pid}")
+                        return True
+                except Exception as e:
+                    logger.error(f"💥 Error pausing process {pid}: {str(e)}")
+            return False
+
+    @staticmethod
+    def resume_process(pid):
+        """Resume a paused process (SIGCONT)"""
+        with process_lock:
+            if pid in active_processes:
+                try:
+                    process_obj = active_processes[pid]["process"]
+                    if process_obj and process_obj.poll() is None:
+                        os.kill(pid, signal.SIGCONT)
+                        active_processes[pid]["status"] = "running"
+                        logger.info(f"▶️  RESUMED: Process {pid}")
+                        return True
+                except Exception as e:
+                    logger.error(f"💥 Error resuming process {pid}: {str(e)}")
+            return False
+
+# Enhanced color codes and visual elements for modern terminal output
+# All color references consolidated to ModernVisualEngine.COLORS for consistency
+    BG_GREEN = '\033[42m'
+    BG_YELLOW = '\033[43m'
+    BG_BLUE = '\033[44m'
+    BG_MAGENTA = '\033[45m'
+    BG_CYAN = '\033[46m'
+    BG_WHITE = '\033[47m'
+
+    # Text effects
+    DIM = '\033[2m'
+    UNDERLINE = '\033[4m'
+    BLINK = '\033[5m'
+    REVERSE = '\033[7m'
+    STRIKETHROUGH = '\033[9m'
